Final code updates
