+++ main.css
@@ -17,7 +17,6 @@
 body {
 	margin: 0;
 	font-family: 'Courier New', Courier, monospace;
-	background-color: var(--main-color);
 	color: var(--text-color);
 }
 
@@ -27,15 +26,24 @@ h1, h2 {
 }
 
 header, main, footer {
+	position: absolute;
+	background-color: rgba(0, 0, 0, 0.9);
 	padding: 10px;
 }
 
 header {
+	top: 0px;
+	width: 100vw;
 	border-bottom: 2px dotted var(--accent-color);
 }
 
+main {
+	top: 200px;
+	margin-left: 20px;
+	border: 2px dotted var(--accent-color);
+}
+
 footer {
-	position: fixed;
 	bottom: 0px;
 	width: 100vw;
 	border-top: 2px dotted var(--accent-color);
